+++ NEWS.md
@@ -2,6 +2,8 @@ Version 0.11.4 For Dev-C++ 7 Beta
  - fix: compiler set's custom link parameters  not used when compiling
  - fix: code completion doesn't work when input inside () or []
  - fix: auto indent processing error when input '{' in the middle of if statement
+ - fix: left and right gutter offset settings not  correctly saved
+ - enhancement: use svg icons for editor gutter, and they can zoom with font now
 
 Version 0.11.3 For Dev-C++ 7 Beta
  - fix: use pixel size for fonts, to fit different dpi in multiple displays